requirements.txt
celery==5.4.0
Django==5.0.8
django-celery-beat==2.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.8
